+++ lib/service/Questionnaire.dart
@@ -1,8 +1,9 @@
 
-import 'package:flutter/material.dart';
-import 'package:flutter/services.dart';
+import 'package:http/http.dart';
 import 'package:illinois/model/Questionnaire.dart';
 import 'package:illinois/service/Auth2.dart';
+import 'package:illinois/service/Config.dart';
+import 'package:rokwire_plugin/service/network.dart';
 import 'package:rokwire_plugin/utils/utils.dart';
 
 class Questionnaires /* with Service */ {
@@ -19,8 +20,13 @@ class Questionnaires /* with Service */ {
   // Implementation
 
   Future<Questionnaire?> loadResearch() async {
-    try { return Questionnaire.fromJson(JsonUtils.decodeMap(await rootBundle.loadString('assets/questionnaire.demographics.json'))); }
-    catch(e) { debugPrint(e.toString()); }
+    if (Config().contentUrl != null) {
+      Response? response = await Network().get("${Config().contentUrl}/content_items", body: JsonUtils.encode({'categories': ['research_questionnaire']}), auth: Auth2());
+      List<dynamic>? responseList = (response?.statusCode == 200) ? JsonUtils.decodeList(response?.body)  : null;
+      dynamic responseItem = ((responseList != null) && responseList.isNotEmpty) ? responseList.first : null;
+      Map<String, dynamic>? responseData = (responseItem is Map) ? JsonUtils.mapValue(responseItem['data']) : null;
+      return Questionnaire.fromJson(responseData);
+    }
     return null;
   }