+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Unreleased
 ### Changed
 - Research Project updates [#2544](https://github.com/rokwire/illinois-app/issues/2544).
+- Load research questionnaire from content service.
 
 ## [4.2.21] - 2022-11-17
 ### Changed